no more sizing warnings in 64 bit mode!

+++ Engine/Source/Bombast/interfaces.h
@@ -116,9 +116,9 @@ class IResourceFile
 public:
 	virtual bool VOpen() = 0;
 	virtual int VGetRawResourceSize(const Resource &r) = 0;
-	virtual int VGetRawResource(const Resource &r, char *buffer) = 0;
-	virtual int VGetNumResources() const = 0;
-	virtual std::string VGetResourceName(int num) const = 0;
+	virtual size_t VGetRawResource(const Resource &r, char *buffer) = 0;
+	virtual size_t VGetNumResources() const = 0;
+	virtual std::string VGetResourceName(size_t num) const = 0;
 	virtual ~IResourceFile() { }
 };
